feat: seperate antei setup into multiple tasks
